requirements.txt
Django==4.2
mysql-connector-python==8.1.0
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0